refactors. delete method implemented
