--- src/main/java/com/alkemy/ong/data/exception/ParamNotFoundException.java
@@ -1,5 +1,0 @@
-package com.alkemy.ong.data.exception;
-
-public class ParamNotFoundException extends RuntimeException{
-    public ParamNotFoundException(String error){super(error);}
-}
